FE: rename folder saved albums
