+++ src/views/my-page/constants/my-page.paths.js
@@ -2,5 +2,5 @@ export const MyPagePaths = {
   MY_PAGE: "/my-page",
   PROFILE: "/profile",
   ALBUM_LIST: "/album-list",
-  SAVED_ALBUMS: "/saved-albums",
+  BOOKMARK_LIST: "/bookmark-list",
 };
